refactor(core-utils): introduce new util file for route stuff
BREAKING CHANGE:

moves util functions related to routes to a new util file
