+++ packages/core-utils/src/index.js
@@ -3,6 +3,7 @@ import * as map from "./map";
 import * as profile from "./profile";
 import * as query from "./query";
 import * as queryParams from "./query-params";
+import * as route from "./route";
 import * as storage from "./storage";
 import * as time from "./time";
 import * as types from "./types";
@@ -14,6 +15,7 @@ const core = {
   profile,
   query,
   queryParams,
+  route,
   storage,
   time,
   types,
